+++ cli/lib/src/functions/functions.dart
@@ -1,7 +1,6 @@
 export 'flutter/is_flutter_project.dart';
 export 'log/logger.dart';
 export 'init/init.dart';
-export 'add/add.dart';
 export 'components_path/components_path.dart';
 export 'close.dart';
 export 'component_methods/component_methods.dart';