+++ fcnui_base/example/lib/main.dart
@@ -1,6 +1,10 @@
 /// 0. Components can be found in: lib/components (or the folder specified in fcnui.json)
 import 'package:example/components/button.dart';
 import 'package:example/components/card.dart';
+import 'package:example/components/form.dart';
+import 'package:example/components/input.dart';
+import 'package:example/components/save_button.dart';
+import 'package:example/components/with_label.dart';
 import 'package:fcnui_base/fcnui_base.dart';
 import 'package:flutter/material.dart';
 
@@ -42,30 +46,112 @@ class _MyHomePageState extends State<MyHomePage> {
         title: Text(widget.title),
       ),
       body: Center(
-        child: Column(
-          mainAxisAlignment: MainAxisAlignment.center,
-          children: <Widget>[
-            /// 2. Use fcnui components ex:[DefaultButton]
-            DefaultButton(
-              variant: PrimaryButtonVariant(
-                onPressed: () {
-                  ChangeFlexSchemeAction(flexScheme: FlexScheme.wasabi.name)
-                      .payload();
-                },
-                text: "Increment",
-                icon: Icons.add,
+        child: SingleChildScrollView(
+          padding: const EdgeInsets.all(20),
+          child: Column(
+            mainAxisAlignment: MainAxisAlignment.center,
+            children: <Widget>[
+              /// 2. Use fcnui components ex:[DefaultButton]
+              DefaultButton(
+                variant: PrimaryButtonVariant(
+                  onPressed: () {
+                    ChangeFlexSchemeAction(flexScheme: FlexScheme.wasabi.name)
+                        .payload();
+                  },
+                  text: "Increment",
+                  icon: Icons.add,
+                ),
               ),
-            ),
 
-            const DefaultCard(
-              variant: CardVariant(
-                  title: CardTitle(title: "Hello"),
-                  content: CardContent(content: Text("Context")),
-                  footer: CardFooter(footer: [])),
-            ),
-          ],
-        ).spaced(20),
+              const DefaultCard(
+                variant: CardVariant(
+                    title: CardTitle(title: "Hello"),
+                    content: CardContent(content: Text("Context")),
+                    footer: CardFooter(footer: [])),
+              ),
+
+              _DecoratedCard(),
+            ],
+          ).spaced(20),
+        ),
       ),
     );
+  }
+}
+
+class _DecoratedCard extends StatelessWidget {
+  _DecoratedCard({super.key});
+
+  final formModel = FormModel();
+
+  @override
+  Widget build(BuildContext context) {
+    return ThemeProvider(builder: (context, vm) {
+      final colorScheme = vm.theme.colorScheme;
+      return DefaultCard(
+        decoration: CardDecoration(
+          borderRadius: BorderRadius.circular(24),
+          border: Border.all(
+              color: colorScheme.primaryContainer,
+              width: 2,
+              strokeAlign: BorderSide.strokeAlignInside),
+          color: colorScheme.primaryContainer.withOpacity(0.2),
+        ),
+        variant: CardVariant(
+          title: const CardTitle(title: "Create project"),
+          subtitle: const CardSubtitle(
+              subtitle: "Deploy your new project in one-click"),
+          content: CardContent(
+            content: DefaultForm(
+              vm: formModel,
+              child: Column(
+                children: [
+                  WithLabel(
+                    labelVm: const LabelModel(text: "Name"),
+                    child: DefaultInput(
+                      vm: InputModel(
+                        name: "name",
+                        validators: [
+                          FormBuilderValidators.required(),
+                        ],
+                        hintText: "Name of the project",
+                      ),
+                    ),
+                  ),
+                  const WithLabel(
+                    labelVm: LabelModel(text: "Description"),
+                    child: DefaultInput(
+                      vm: InputModel(
+                        name: "description",
+                        hintText: "Description of the project",
+                      ),
+                    ),
+                  ),
+                ],
+              ).spaced(20),
+            ),
+          ),
+          footer: CardFooter(
+            footer: [
+              DefaultButton(
+                variant: OutlineButtonVariant(
+                  onPressed: () {},
+                  text: "Cancel",
+                ),
+              ),
+              SaveButton(
+                  vm: formModel,
+                  text: "Deploy",
+                  onSave: (value) {
+                    if (formModel.isValid) {
+                      ScaffoldMessenger.of(context).showSnackBar(SnackBar(
+                          content: Text(formModel.getValues().toString())));
+                    }
+                  })
+            ],
+          ),
+        ),
+      );
+    });
   }
 }